+++ lib/rubocop/rspec/language/each_selector.rb
@@ -1,0 +1,13 @@
+# Monkey-patch SelectorSet to allow enumeration of selectors.
+
+module RuboCop
+  module RSpec
+    module Language
+      class SelectorSet
+        def each(&blk)
+          selectors.each(&blk)
+        end
+      end
+    end
+  end
+end